% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatedPlot.R
\name{animatedPlot}
\alias{animatedPlot}
\title{Creates an animated plot function}
\usage{
animatedPlot(plotFun, plotsPerSecond = 5, adapt = T)
}
\arguments{
\item{plotFun}{a plotting function}

\item{plotsPerSecond}{maximum number of plots per second}

\item{adapt}{if \code{TRUE}, adapt the plotting interval to ensure plotting
takes no more than half of all processing time.}
}
\value{
a animated plotting function
}
\description{
Create an animated plot function from a plot function. Pass a function which
generates your plots (eg \link{plot}, \link{hist}) and the maximum number of
plots per second and this function returns a function which will ensure the
function plots no more than \code{plotsPerSecond} times per second.
}
\details{
You then call this function instead of your original function, passing the
same parameters you would pass to the original.
}
\examples{
aplot <- animatedPlot(plot, 5)
walk <- 0
xlim <- c(0,10000)
for (i in 1:9999) {
  Sys.sleep(0.001)
  walk <- c(walk, walk[i] + runif(1) - 0.5)
  aplot(seq_along(walk), walk, type="l", xlim=xlim)
}
}

