% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingWindow.R
\name{loggingWindow}
\alias{loggingWindow}
\title{Create a logging window using TCL/TK}
\usage{
loggingWindow(height = 30, width = 100, timestamp = TRUE)
}
\arguments{
\item{timestamp}{}
}
\description{
Create a logging window using TCL/TK
}
\examples{
win <- loggingWindow()
updateLog(win, "Hello World!")
}

