% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTkProgressBarThrottled.R
\name{setTkProgressBarThrottled}
\alias{setTkProgressBarThrottled}
\title{A Tk Progress Bar with Throttled Updates}
\usage{
setTkProgressBarThrottled(pb, value, title = NULL, label = NULL)
}
\arguments{
\item{updateFreq}{}
}
\description{
A Tk Progress Bar with Throttled Updates
}
\examples{
pb <- tkProgressBar("Example", min=0, max=300, initial=0)
for (i in 1:300) {
  Sys.sleep(0.1)
  setTkProgressBarThrottled(pb, i)
}
close(pb)
}

