% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkProgressBarThrottled.R
\name{tkProgressBarThrottled}
\alias{tkProgressBarThrottled}
\title{Create a throttled Tk progress bar}
\usage{
tkProgressBarThrottled(title = "R progress bar", label = "", min = 0,
  max = 1, initial = 0, width = 300, updateFreq = 1)
}
\arguments{
\item{updateFreq}{}
}
\description{
Create a throttled Tk progress bar
}

