% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingWindow.R
\name{updateLog}
\alias{updateLog}
\title{Add a line to a logging window}
\usage{
updateLog(win, message)
}
\arguments{
\item{message}{}
}
\description{
Add a line to a logging window
}
\examples{
win <- loggingWindow()
updateLog(win, "Hello World!")
}

