% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTkProgressBarThrottled.R
\name{closeTkProgressBarThrottled}
\alias{closeTkProgressBarThrottled}
\title{Close a throttled Tk progress bar}
\usage{
closeTkProgressBarThrottled(pb)
}
\arguments{
\item{pb}{}
}
\description{
Close a throttled Tk progress bar
}

